Refactored Poller example to separate package to limit @ComponentScan scope.
