+++ annotations/basic/helloworld/src/main/resources/log4j2.xml
@@ -10,7 +10,7 @@
 			<AppenderRef ref="Console" />
 		</Root>
 		<Logger name="org.springframework" level="warn" />
-		<Logger name="org.springframework.integration" level="info" />
+		<Logger name="org.springframework.integration" level="warn" />
 		<Logger name="org.springframework.integration.samples" level="debug" />
 	</Loggers>
 </Configuration>
